refactor: wave11 peels — chat, projects, scheduling, event-source
Behavior-preserving package organization after wave 10:
- app/api/event-source.ts (resilient SSE reconnect)
- app/api/chat, projects, agents, workflows, scheduling, ai-text, ai-sessions, research
- rename remaining-ops-6 → task-store/branch-and-pr-entities.ts
Ratchet legacy.ts line-count ceiling downward (~8921 → ~5670).
